Directly specify monster xp values
Monster experience values relied on a very complicated formula that tried to
account for a large number of factors related to monster power. While it did
capture many of the aspects that make a monster more dangerous, it was
inaccurate enough that monsters needed an additional corrective factor to have
reasonable values. Since we're already specifying monster xp with an ad hoc
formula, it's simpler to just remove the formula and specify xp value directly.

This commit does that for almost all monsters. Certain derived or highly
variable monsters (like panlords) will be handled separately.

+++ crawl-ref/source/mon-util.h
@@ -131,7 +131,7 @@ struct monsterentry
     resists_t resists;
 
     // Multiplier for calculated monster XP value; see exper_value() for use.
-    int8_t exp_mod;
+    int exp_mod;
 
     monster_type genus,         // "team" the monster plays for
                  species;       // corpse type of the monster
